Clean up

+++ SDwebsite/database/migrations/2021_02_26_223712_fuel_quote_history_table.php
@@ -22,8 +22,8 @@ class FuelQuoteHistoryTable extends Migration
             $t -> float('Gallons', 200)->nullable();
             $t -> string('Address', 500)->nullable();
             $t -> timestamp('start')->nullable();
-            $t -> float('Suggested_Price')->nullable();
-            $t -> float('Due')->nullable();
+            $t -> double('Suggested_Price')->nullable();
+            $t -> double('Due')->nullable();
         });
 
     }
